+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 Note the first digit of every adapter version corresponds to the major version of the Chartboost Mediation SDK compatible with that adapter. 
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
+### 5.23.2.0.0
+- This version of the adapter has been certified with Google Mobile Ads SDK 23.2.0.
+- This version of the adapter supports Chartboost Mediation SDK version 5.+.
+
 ### 4.22.3.0.6
 - Initialization of `MobileAds` SDK on the `IO` context.